requirements.txt
Flask==0.12.2
gunicorn==19.6.0
line-bot-sdk==1.3.0
imgurpython==1.1.7
numpy
opencv-python==3.2.0.8
cmake
dlib
matplotlib
scipy
scikit-image
rq==0.7.1
pytz==2017.2
imutils
Pillow
